Final Header fix
Bench: 2401193

+++ src/history.cpp
@@ -2,7 +2,7 @@
 
 int DON::loop() {
     int k = 2;
-    for (size_t i = 0; i < 10; ++i)
+    for (int i = 0; i < 10; ++i)
     {
         ++k;
     }
